pin dependencies in conda environment YAML, splist transdecoder and TD2 tests into separate test modules, compare transdecoder and td2 to separate reference outputs, and alter README section to only suggest building the conda environment from the YAML file
