+++ test/data/output_transdecoder/quickprot.cds.fasta
