+++ test/data/output_transdecoder/quickprot.longest.cds.fasta
