+++ test/data/output_transdecoder/quickprot.longest.pep.fasta
