+++ test/data/output_transdecoder/quickprot.pep.fasta
