+++ test/data/output_transdecoder/quickprot.transcript.gtf
