+++ test/data/output_transdecoder/quickprot.uniprotkb_reviewed_true_AND_taxonomy_id_4930_Saccharomyces.fasta.miniprot_output.outs_0.95.gff3
